backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-contrib-python
Pillow
numpy
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv